+++ packages/vda-sbt-client/src/blockchain/blockchainApi.ts
@@ -2,38 +2,18 @@ import {
     getVeridaContract,
     VeridaContract
 } from "@verida/web3"
-import {
-    Web3SelfTransactionConfig,
-    Web3MetaTransactionConfig,
-    Web3CallType,
-    EnvironmentType
-} from '@verida/types'
+import { Web3SelfTransactionConfig, VdaClientConfig } from '@verida/types'
 import { getContractInfoForNetwork, RPC_URLS } from "@verida/vda-common";
 import { JsonRpcProvider } from '@ethersproject/providers';
 import { explodeDID } from '@verida/helpers'
 
-import { ethers, Contract } from "ethers";
+import { ethers, Contract, BigNumberish } from "ethers";
 import EncryptionUtils from "@verida/encryption-utils";
 
-/**
- * Interface for vda-sbt-client instance creation. Same as VDA-DID configuration
- * @param did @string DID
- * @param signKey @string Private key of DID (hex string). Used to generate signature in transactions to chains
- * @param network @string Target chain name or chain id.
- * @param callType @string 'web3' | 'gasless'
- * @param web3Options object Web3 configuration depending on call type. Same as vda-did-resolver
- */
-export interface SBTClientConfig {
-    network: EnvironmentType
-    did?: string
-    signKey?: string
-    callType?: Web3CallType
-    web3Options?: Web3SelfTransactionConfig | Web3MetaTransactionConfig
-}
 
 export class VeridaSBTClient {
 
-    private config: SBTClientConfig
+    private config: VdaClientConfig
     private network: string
     private didAddress?: string
 
@@ -41,7 +21,7 @@ export class VeridaSBTClient {
     private vdaWeb3Client?: VeridaContract
     private contract?: ethers.Contract
 
-    public constructor(config: SBTClientConfig) {
+    public constructor(config: VdaClientConfig) {
         if (!config.callType) {
             config.callType = 'web3'
         }
@@ -101,7 +81,7 @@ export class VeridaSBTClient {
      * @param tokenId tokenId
      * @returns tokenURI from SBT contract
      */
-    public async tokenURI(tokenId: number) {
+    public async tokenURI(tokenId: BigNumberish) {
         let response
         try {
             if (this.vdaWeb3Client) {
@@ -127,7 +107,7 @@ export class VeridaSBTClient {
      * @param tokenId Token ID
      * @returns true if tokenID is locked
      */
-    public async isLocked(tokenId: number) {
+    public async isLocked(tokenId: BigNumberish) {
         let response
         try {
             if (this.vdaWeb3Client) {
@@ -340,7 +320,7 @@ export class VeridaSBTClient {
      * @returns string array of SBT type & uniqueId
      */
     public async tokenInfo(
-        tokenId: number
+        tokenId: BigNumberish
     ) {
         let response
         try {
@@ -369,7 +349,7 @@ export class VeridaSBTClient {
      * @param tokenId SBT tokenId
      */
     public async burnSBT(
-        tokenId: number
+        tokenId: BigNumberish
     ) {
         if (this.readOnly || !this.config.signKey) {
             throw new Error(`Unable to submit to blockchain. In read only mode.`)
@@ -388,7 +368,7 @@ export class VeridaSBTClient {
      * @returns owner address of the token
      */
     public async ownerOf(
-        tokenId: number
+        tokenId: BigNumberish
     ) {
         let response
         try {